refactor: abort if id is nil

+++ lib/rujira/api/issue.rb
@@ -15,6 +15,7 @@ module Rujira
       end
 
       def get(id_or_key, &block)
+        abort 'Issue ID or KEY is required' if id_or_key.nil?
         builder do
           path "issue/#{id_or_key}"
           instance_eval(&block) if block_given?
@@ -23,6 +24,7 @@ module Rujira
       end
 
       def delete(id_or_key, &block)
+        abort 'Issue ID or KEY is required' if id_or_key.nil?
         builder do
           path "issue/#{id_or_key}"
           method :delete
@@ -34,6 +36,7 @@ module Rujira
       alias del delete
 
       def edit(id_or_key, &block)
+        abort 'Issue ID or KEY is required' if id_or_key.nil?
         builder do
           path "issue/#{id_or_key}"
           method :put
@@ -43,10 +46,12 @@ module Rujira
       end
 
       def comment(id_or_key, &block)
+        abort 'Issue ID or KEY is required' if id_or_key.nil?
         @client.Comment.create id_or_key, &block
       end
 
       def watchers(id_or_key, name, &block)
+        abort 'Issue ID or KEY is required' if id_or_key.nil?
         builder do
           path "issue/#{id_or_key}/watchers"
           method :post
@@ -57,6 +62,7 @@ module Rujira
       end
 
       def attachments(id_or_key, path, &block)
+        abort 'Issue ID or KEY is required' if id_or_key.nil?
         @client.Attachments.create id_or_key, path, &block
       end
     end
